SERVER-69527 add gating logic to fle2 range tests
GitOrigin-RevId: 5eff54799ae8edb4e135fbd1b7b3d567e7579bb4

+++ src/mongo/db/modules/enterprise/jstests/fle2/txn_insert_range.js
@@ -12,6 +12,12 @@ load("jstests/fle2/libs/encrypted_client_util.js");
 (function() {
 'use strict';
 
+// TODO SERVER-67760 remove once feature flag is gone
+if (!isFLE2RangeEnabled()) {
+    jsTest.log("Test skipped because featureFlagFLE2Range is not enabled");
+    return;
+}
+
 let dbName = 'txn_insert_range';
 let dbTest = db.getSiblingDB(dbName);
 dbTest.dropDatabase();
